# postgresql-backend-2-server/swagger.yaml
openapi: 3.0.0
info:
  title: Furniture Store API
  version: 1.0.0
  description: API documentation for the furniture store backend (PostgreSQL + Sequelize)
servers:
  - url: http://localhost:3000/api
paths:
...
  /products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
    patch:
      summary: Update product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductPatch'
      responses:
        '204':
          description: Product updated
        '400':
          description: Bad request
    delete:
      summary: Delete product (admin only)
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
  /products:
    get:
      summary: Get product by slug
      parameters:
        - in: query
          name: slug
          schema:
            type: string
      responses:
        '200':
          description: Product found or empty list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        namn:
          type: string
        description:
          type: string
        price:
          type: integer
        image:
          type: string
        slug:
          type: string
...
    ProductPatch:
      type: object
      properties:
        op:
          type: string
        path:
          type: string
        value:
          type: string
...